- Create util for checking release date expiration - Apply the util on sidebar
